incorporated comments from code review
